requirements.txt
streamlit 
pyautogen==0.2.28
google-generativeai
chromadb== 0.5.0
markdownify==0.12.1
ipython==8.24.0
streamlit==1.35.0
pysqlite3-binary
sentence_transformers==2.7.0